uploads: Move unauth-signed tokens into view.
